Refactor User Catches Map
Make mobile reponsive
Clean up code
