+++ client/src/pages/UserCatchesMap/UserCatchesMap.css
@@ -1,0 +1,90 @@
+.map-container {
+ position: relative;
+ height: 100%;
+ width: 100%;
+ min-width: 350px;
+}
+
+.autocomplete-input {
+  /* position: absolute;
+  top: 10px;
+  left: 50%;
+  transform: translateX(-50%); */
+  width: 250px;
+  height: 35px;
+  margin: 0;
+  border-radius: 5;
+  z-index: 3;
+  margin: 5px;
+}
+
+.filter-menu {
+  margin: 5px
+  /* top: 50px;
+  left: calc(50% - 100px); */
+}
+
+.filters-and-autocomplete-container {
+  display: flex;
+  position: absolute;
+  top: 0px;
+  left: 50%;
+  transform: translateX(-50%);
+  z-index: 3
+}
+
+@media screen and (max-width: 600px) {
+  .filters-and-autocomplete-container {
+    flex-direction: column-reverse;
+    left: 0px;
+    transform: none;
+  }
+}
+
+
+.catch-cards-container {
+  padding: 0px 10px;
+  margin-top: 20px;
+  width: 400px;
+  height: 100%;
+  overflow-y: auto;
+}
+
+@media screen and (max-width: 960px) {
+  .catch-cards-container {
+    display: none;
+  }
+}
+
+.log-catch-map-buttons-container {
+  display: flex;
+  flex-wrap: wrap-reverse;
+  justify-content: center;
+  align-items: center;
+  position: absolute;
+  bottom: 30px;
+  left: 50%;
+  transform: translateX(-50%);
+}
+
+.log-catch-map-button {
+  padding: 0px;
+  margin: 5px !important;
+  height: 40px;
+  width: 150px; 
+  z-index: 3; 
+}
+
+.catch-form-container {
+  margin: 30px auto;
+  width: 80%;
+  max-width: 400px;
+  min-width: 300px;
+}
+
+.catch-date {
+  position: absolute;
+  top: 0px;
+  left: 100%;
+  min-width: 175px;
+}
